Updated papckaging approach for reducing the number of github requests This PR is needed only for huge updates like year switch in header files.
Currently per 1 Asset Upload-Delete request function call we use 4-20 github requests.
With 450 assets (at the moment when this PR was created) we exceed 5000 request limit if we need to reupload all of them (we always have chack indexes script running in the background twice in an hour and 1 scheduled workflow at the moment takes 900 requests)

So with this update packaging script will take ~10 requests in the beginning to take all uploaded assets and then just 2 requests for Asset Delete-Upload function call.